First pass at header only library.
