+++ include/jmdlx.h
@@ -14,10 +14,10 @@
 #define JMD_DLX_NAMESPACE_END }}
 
 #include <string>
+#include <sstream>
 #include <vector>
 #include <set>
 #include <list>
-
 #include <stdexcept>
 #include <climits> // LONG_MAX
 
@@ -38,7 +38,13 @@ class matrix_one
     matrix_one() : left(this), right(this), up(this), down(this),
                    hdr_ptr(this), rc_idx(-1) {}
     // inserting constructor: creates and inserts at the specified location
-    matrix_one(intz_t rc_idx, matrix_one* hdr_ptr, matrix_one* prev);
+    matrix_one(intz_t rc_idx, matrix_one* hdr_ptr, matrix_one* prev)
+            : left(prev ? prev : this), right(prev ? prev->right : this),
+              up(hdr_ptr?hdr_ptr->up:this), down(hdr_ptr?hdr_ptr:this),
+              hdr_ptr(hdr_ptr), rc_idx(rc_idx)
+    {
+        up->down = down->up = right->left = left->right = this;
+    }
 
     void column_unlink() {
         right->left = left;
@@ -98,17 +104,62 @@ public:
 class ec_matrix
 {
 public:
-    ec_matrix(intz_t col_count, const all_rows& rows, ec_callback& eccb_);
-    ~ec_matrix();
+    ec_matrix(intz_t col_count, const all_rows& rows, ec_callback& eccb)
+            : quit_searching_(false), col_specs_(col_count), eccb_(eccb)
+    {
+        intz_t i;
+        matrix_one* prev = &root_;
+        for (i= 0; i<col_count; ++i) {
+            matrix_one* next = new matrix_one(i, NULL, prev);
+            heap_ones_.push_back(next);
+            col_specs_[i].hdr_ptr = next;
+            prev = next;
+        }
+
+        intz_t row_count = rows.size();
+        for (i = 0; i<row_count; ++i) { // for each row in all_rows
+            const row_spec& row = rows[i];
+            bool constraint = row.constraint;
+            matrix_one* prev = NULL;
+            std::list<intz_t>::const_iterator it, endit=row.col_indices.end();
+            for (it=row.col_indices.begin(); it!=endit; ++it) {
+                // for each column index in which there is a 1 in this row
+                intz_t col_idx = *it;
+                prev = new matrix_one(i, col_specs_[col_idx].hdr_ptr, prev);
+                heap_ones_.push_back(prev);
+                if (constraint) {
+                    std::pair<ones_set::iterator, bool> inserted =
+                        constraint_hdrs_.insert(prev->hdr_ptr);
+                    if (!inserted.second) {
+                        // sudoku's not one, but is there a use case where this is
+                        // not an error?
+                        std::ostringstream os;
+                        os << "Overconstrained on column " << col_idx;
+                        throw std::runtime_error(os.str());
+                    }
+                }
+                ++col_specs_[col_idx].size;
+            }
+        }
+        try {
+            prune_constraints();
+        }
+        catch(std::exception& e) {
+            cleanup();
+            throw e;
+        }
+    }
+
+    ~ec_matrix()
+    {
+        cleanup();
+    }
 
     void search() {
         search(0);
     }
 
 protected:
-    void prune_constraints();
-    void cleanup();
-
     matrix_one* best_column() {
         intz_t min_ones = INTZ_MAX;
         matrix_one* best = NULL;
@@ -180,6 +231,21 @@ protected:
             }
             uncover_column(best_hdr);
         }
+    }
+
+    void prune_constraints()
+    {
+        ones_set::iterator it, endit=constraint_hdrs_.end();
+        for (it=constraint_hdrs_.begin(); it!=endit; ++it) {
+            cover_column(*it);
+        }
+    }
+
+    void cleanup()
+    {
+        std::list<matrix_one*>::const_iterator it, endit=heap_ones_.end();
+        for (it=heap_ones_.begin(); it!=endit; ++it)
+            delete *it;
     }
 
     matrix_one root_;
