--- lib/Makefile.am
@@ -1,7 +1,0 @@
-# $Id: Makefile.am,v 1.1.1.1 2008/04/09 20:40:19 mark Exp $
-
-libjmdlx_la_CPPFLAGS = -I$(top_builddir)/include
-METASOURCES = AUTO
-lib_LTLIBRARIES = libjmdlx.la
-libjmdlx_la_SOURCES = jmdlx.cpp
-libjmdlx_la_LDFLAGS = -version-info 0:1:0
