+++ src/Makefile.am
@@ -9,8 +9,6 @@ sudoku_dlx_CPPFLAGS= -I$(top_builddir)/include $(all_includes)
 # the library search path.
 sudoku_dlx_LDFLAGS = $(all_libraries) 
 
-sudoku_dlx_LDADD = ../lib/libjmdlx.la
-
 ## sudoku_dlx_CXXFLAGS = -DCOLUMN_SETUP_DEBUG=1
 ## sudoku_dlx_CXXFLAGS = -DROW_SETUP_DEBUG=1
 ## sudoku_dlx_CXXFLAGS = -DCOLUMN_SETUP_DEBUG=1 -DROW_SETUP_DEBUG=1
