+++ src/sudoku_main.cpp
@@ -8,10 +8,6 @@
  *   https://github.com/jmderic/sudoku_dlx                                 *
  ***************************************************************************/
 
-#ifdef HAVE_CONFIG_H
-#include <config.h>
-#endif
-
 #include <iostream>
 #include <cstdlib>
 #include <algorithm>
